backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
pydantic[email]
typing-extensions
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]